pyproject.toml
[project]
name = "ttv"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pillow>=12.2.0",
    "pygments>=2.20.0",
]
